requirements.txt
selenium
requests
discord-py